Reformatted, added block to changeLocalDir test to only run if connection

+++ src/test/java/ClientTest.java
@@ -88,7 +88,6 @@ public class ClientTest {
     Client client = new Client(password, hostName, userName);
     client.connect();
 
-    client.deleteRemoteFile(fileName);
     client.uploadFile(fileName);
     attrs = client.getcSftp().stat(fileName);
     if (attrs != null)
@@ -221,28 +220,29 @@ public class ClientTest {
     File newDir = new File(newLocalPath);
 
     Client client = new Client(password, hostName, userName);
-    client.connect();
+    if (client.connect()) {
 
-    if (newDir.mkdir()) {          //create new directory path
-      System.setOut(new PrintStream(output));
-      output.reset();
-      client.printLocalWorkingDir();
-      assertThat(output.toString().contains(newLocalPath), equalTo(false)); //assert current path is not newDir
-      client.changeLocalWorkingDir(newLocalPath);       //change path to newDir
-      output.reset();
-      client.printLocalWorkingDir();
-      assertThat(output.toString(), containsString(newLocalPath));    //assert current path is newDir
-      client.changeLocalWorkingDir("..");       //reset path
-      System.setOut(stdout);    //reset output to standard System.out
-      if (!newDir.delete())
-        System.out.println("Error deleting testing directory");
-      else {
-        System.out.println("Path successfully changed to new dir. New dir has been deleted and path is reset.");
-        pass = true;
+      if (newDir.mkdir()) {          //create new directory path
+        System.setOut(new PrintStream(output));
+        output.reset();
+        client.printLocalWorkingDir();
+        assertThat(output.toString().contains(newLocalPath), equalTo(false)); //assert current path is not newDir
+        client.changeLocalWorkingDir(newLocalPath);       //change path to newDir
+        output.reset();
+        client.printLocalWorkingDir();
+        assertThat(output.toString(), containsString(newLocalPath));    //assert current path is newDir
+        client.changeLocalWorkingDir("..");       //reset path
+        System.setOut(stdout);    //reset output to standard System.out
+        if (!newDir.delete())
+          System.out.println("Error deleting testing directory");
+        else {
+          System.out.println("Path successfully changed to new dir. New dir has been deleted and path is reset.");
+          pass = true;
+        }
+      } else {
+        System.setOut(stdout);
+        System.out.println("Error in mkdir");
       }
-    } else {
-      System.setOut(stdout);
-      System.out.println("Error in mkdir");
     }
     assertThat(pass, equalTo(true));
   }
@@ -322,10 +322,9 @@ public class ClientTest {
       System.setOut(stdout);      //Reset printstream to System.out
       System.out.println("New remote file successfully displayed. New dir has been deleted.");
       pass = true;
-    } 
-    else 
+    } else
       System.out.println("Error in createRemoteDir");
-    
+
     assertThat(pass, equalTo(true));
   }
 }
